requirements.txt
Flask==2.3.2
torch
torchvision
Pillow
paddleocr
opencv-python
